Modify terminal properties by interface
The existing terminal helper methods in pkg/term operate on file
descriptors, but some Windows terminal properties are not controllable
through file descriptor when VT100 emulation is provided by go-ansiterm.
This change update the terminal helpers and all callers to operate on a
terminal interface so that a future change can improve Windows VT100
handling.

Unfortunately, the vendored notary package has a dependency on the
existing helpers, so this change keeps a few of the old helpers around
until notary can be updated.

Signed-off-by: John Starks <jostarks@microsoft.com>